requirements.txt
asgiref==3.8.1
Django==4.2
django-background-tasks==1.2.8
django-jalali==7.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
packaging==25.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
